PLT-229 Fixed resuming of UTXO indexer
* Removed 'BlockNo' from the database stored on disk

* Removed the 'DROP TABLES' in the 'open' functions so that we can
  actually be able to resume at startup

* Fixed the UtxoEvent generator to generate unique utxos and only spend utxos which have been
  previously created.

* Added property tests for the resuming capabilities
