+++ marconi-chain-index/changelog.d/20230213_200010_konstantinos.lambrou_PLT_229_fix_resuming_of_utxo_indexer.md
@@ -1,0 +1,10 @@
+### Changed
+
+- `UtxoRow` doesn't store the `BlockNo` anymore
+
+### Fixed
+
+- Fixed the resuming of the UTXO indexer
+- Fixed the UtxoEvent generator to generate unique utxos and only spend utxos which have been
+  previously created.
+
